--- apps/container/src/components/WppDashboard/index.ts
@@ -1,1 +1,0 @@
-export { default as WppDashboard } from './WppDashboard.vue';
